Migrate locking statics to spatial statics, add function details
