--- SpatialGDK/Source/SpatialGDK/Private/Utils/LockingStatics.cpp
@@ -1,81 +1,0 @@
-// Copyright (c) Improbable Worlds Ltd, All Rights Reserved
-
-#include "Utils/LockingStatics.h"
-
-#include "EngineClasses/SpatialNetDriver.h"
-#include "LoadBalancing/AbstractLockingPolicy.h"
-#include "SpatialConstants.h"
-#include "Utils/SpatialStatics.h"
-
-#include "Containers/UnrealString.h"
-
-DEFINE_LOG_CATEGORY(LogLocking);
-
-namespace
-{
-bool CanProcessActor(const AActor* Actor)
-{
-	if (Actor == nullptr)
-	{
-		UE_LOG(LogLocking, Error, TEXT("Calling locking API functions on nullptr Actor is invalid."));
-		return false;
-	}
-
-	const UNetDriver* NetDriver = Actor->GetWorld()->GetNetDriver();
-	if (!NetDriver->IsServer())
-	{
-		UE_LOG(LogLocking, Error, TEXT("Calling locking API functions on a client is invalid. Actor: %s"), *GetNameSafe(Actor));
-		return false;
-	}
-
-	if (!Actor->HasAuthority())
-	{
-		UE_LOG(LogLocking, Error, TEXT("Calling locking API functions on a non-auth Actor is invalid. Actor: %s."),
-			*GetNameSafe(Actor));
-		return false;
-	}
-
-	return true;
-}
-} // anonymous namespace
-
-FLockingToken ULockingStatics::AcquireLock(AActor* Actor, const FString& DebugString)
-{
-	if (!CanProcessActor(Actor) || !USpatialStatics::IsSpatialMultiWorkerEnabled(Actor->GetWorld()))
-	{
-		return FLockingToken{ SpatialConstants::INVALID_ACTOR_LOCK_TOKEN };
-	}
-
-	UAbstractLockingPolicy* LockingPolicy = Cast<USpatialNetDriver>(Actor->GetWorld()->GetNetDriver())->LockingPolicy;
-
-	const ActorLockToken LockToken = LockingPolicy->AcquireLock(Actor, DebugString);
-
-	UE_LOG(LogLocking, Verbose, TEXT("LockingComponent called AcquireLock. Actor: %s. Token: %lld. New lock count: %d"),
-		*Actor->GetName(), LockToken, LockingPolicy->GetActorLockCount(Actor));
-
-	return FLockingToken{ LockToken };
-}
-
-bool ULockingStatics::IsLocked(const AActor* Actor)
-{
-	if (!CanProcessActor(Actor) || !USpatialStatics::IsSpatialMultiWorkerEnabled(Actor->GetWorld()))
-	{
-		return false;
-	}
-
-	return Cast<USpatialNetDriver>(Actor->GetWorld()->GetNetDriver())->LockingPolicy->IsLocked(Actor);
-}
-
-void ULockingStatics::ReleaseLock(const AActor* Actor, FLockingToken LockToken)
-{
-	if (!CanProcessActor(Actor) || !USpatialStatics::IsSpatialMultiWorkerEnabled(Actor->GetWorld()))
-	{
-		return;
-	}
-
-	UAbstractLockingPolicy* LockingPolicy = Cast<USpatialNetDriver>(Actor->GetWorld()->GetNetDriver())->LockingPolicy;
-	LockingPolicy->ReleaseLock(LockToken.Token);
-
-	UE_LOG(LogLocking, Verbose, TEXT("LockingComponent called ReleaseLock. Actor: %s. Token: %lld. Resulting lock count: %d"),
-		*Actor->GetName(), LockToken.Token, LockingPolicy->GetActorLockCount(Actor));
-}
